Логика приложения отделена от API
