+++ app/posts/post_logic.py
@@ -1,14 +1,11 @@
 from typing import List
 
-from app.database.db import db
-from fastapi import APIRouter, HTTPException, status
-from app.models.post import Post, PostCreate
-
-router = APIRouter()
+from database.db import db
+from fastapi import HTTPException, status
+from models.post import Post, PostCreate
 
 
-@router.get("/")
-async def all() -> List[Post]:
+def get_all() -> List[Post]:
     """Return list of all posts.
 
     Returns:
@@ -17,8 +14,7 @@ async def all() -> List[Post]:
     return list(db.posts.values())
 
 
-@router.get("/{id}")
-async def get(id: int) -> Post:
+def get_post(id: int) -> Post:
     """Return post by requested id.
 
     Args:
@@ -36,8 +32,7 @@ async def get(id: int) -> Post:
         raise HTTPException(status.HTTP_404_NOT_FOUND)
 
 
-@router.post("/", status_code=status.HTTP_201_CREATED)
-async def create(post_create: PostCreate) -> Post:
+def create_post(post_create: PostCreate) -> Post:
     """Create a new post.
 
     Args:
@@ -63,8 +58,7 @@ async def create(post_create: PostCreate) -> Post:
     return post
 
 
-@router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete(id: int) -> None:
+def delete_post(id: int) -> None:
     """Deleting a post.
 
     Args:
